requirements.txt
Deprecated==1.2.13
packaging==21.3
pycryptodome==3.12.0
pyparsing==3.0.7
redis==4.1.1
wrapt==1.13.3
